moved separated auth middleware from core middlewares
